dcp-ops-448 Update delete script to use latest ingest client version
