--- scripts/analysis-submission-cleanup/requirements.txt
@@ -1,2 +1,0 @@
-requests
--e git+https://github.com/ebi-ait/ingest-client.git@a7b6e4b#egg=hca_ingest
