+++ scripts/delete_submissions/README.md
@@ -1,0 +1,11 @@
+# Delete submissions
+
+## How to run the script
+1. Initialise environment vars and run the script
+```bash
+export GOOGLE_APPLICATION_CREDENTIALS=<filepath>;
+export INGEST_API=<ingest-api>
+
+python delete_submissions.py filepath
+```
+
